+++ src/Dynatrace.OpenKit/Protocol/ProtocolConstants.cs
@@ -19,7 +19,7 @@ namespace Dynatrace.OpenKit.Protocol
     public static class ProtocolConstants
     {
         // version constants
-        public const string OpenKitVersion = "8.281.30300";
+        public const string OpenKitVersion = "8.297.30300";
         public const int ProtocolVersion = 3;
         public const int PlatformTypeOpenKit = 1;
         public const string AgentTechnologyType = "okdotnet";